Move dashboard views to dashboard app
