+++ amy/workshops/views.py
@@ -103,7 +103,6 @@ from workshops.forms import (
     PersonsMergeForm,
     PersonCreateForm,
     SponsorshipForm,
-    AutoUpdateProfileForm,
     DCSelfOrganizedEventRequestFormNoCaptcha,
     TrainingProgressForm,
     BulkAddTrainingProgressForm,
@@ -111,7 +110,6 @@ from workshops.forms import (
     TrainingRequestUpdateForm,
     TrainingRequestsSelectionForm,
     TrainingRequestsMergeForm,
-    SendHomeworkForm,
     BulkDiscardProgressesForm,
     BootstrapHelper,
     bootstrap_helper_inline_formsets,
